Attempt use of IY.
