+++ preprocessor/Map Preprocessor/RegisterAllocators/Prioritiser.h
@@ -55,20 +55,43 @@ public:
 		return back->first + 1;
 	}
 
+	std::optional<int> priority_at(Time time, Time horizon, IntT value) const {
+		IntT throwaway;
+		const auto priorities = all_priorities(time, horizon, throwaway);
+		const auto entry = priorities.find(value);
+		if(entry == priorities.end()) {
+			return {};
+		}
+		return entry->second.usages_remaining;
+	}
+	
 	std::optional<PrioritisedValue<IntT>> prioritised_value_at(Time time, Time horizon) const {
 		// Version 1 has a very simple metric: number of remaining usages.
 		// TODO: incorporate a sense of how imminent those usages are.
 
 		// Count remaining usages, keeping track of the top item.
 		IntT top_value = 0;
-		size_t top_count = 0;
+		const auto priorities = all_priorities(time, horizon, top_value);
+
+		// Pick whatever came out on top, if anything.
+		if(priorities.empty()) {
+			return {};
+		}
+		return priorities.find(top_value)->second;
+	}
+
+private:
+	std::map<Time, IntT> values_;
+	
+	std::unordered_map<IntT, PrioritisedValue<IntT>> all_priorities(Time time, Time horizon, IntT &top_value) const {
 		std::unordered_map<IntT, PrioritisedValue<IntT>> priorities;
 
+		size_t top_count = 0;
 		auto cursor = values_.lower_bound(time);
 		const auto target = values_.upper_bound(horizon);
 		while(cursor != target) {
 			auto &priority = priorities[cursor->second];
-			
+
 			priority.value = cursor->second;
 			priority.active_range.begin = std::min(priority.active_range.begin, cursor->first);
 			priority.active_range.end = std::max(priority.active_range.end, cursor->first + 1);
@@ -81,14 +104,8 @@ public:
 			++cursor;
 		}
 
-		// Pick whatever came out on top, if anything.
-		if(priorities.empty()) {
-			return {};
-		}
-		return priorities[top_value];
+		return priorities;
 	}
 
-private:
-	std::map<Time, IntT> values_;
 };
